+++ docs/changelog.rst
@@ -4,6 +4,7 @@ v0.4dev (XX/XX/XX)
 ------------------
 New features
 ^^^^^^^^^^^^
+- add the name of the line with maximum SNR in ztable
 - line_plot function 
 - the bounds for EMCEE fit are reduced and set with respect to the LSQ first fit
 - improve unit tests